+++ packages/carp_context_package/lib/context.dart
@@ -8,6 +8,7 @@ library context;
 import 'dart:async';
 import 'dart:math' as math;
 import 'package:carp_background_location/carp_background_location.dart';
+import 'package:flutter/material.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 import 'package:activity_recognition_flutter/activity_recognition_flutter.dart';