+++ packages/carp_context_package/test/carp_context_package_test.dart
@@ -5,7 +5,8 @@ import 'package:carp_context_package/context.dart';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 import 'package:test/test.dart';
 
-String _encode(Object object) => const JsonEncoder.withIndent(' ').convert(object);
+String _encode(Object object) =>
+    const JsonEncoder.withIndent(' ').convert(object);
 
 void main() {
   Study study;
@@ -13,10 +14,15 @@ void main() {
   setUp(() {
     SamplingPackageRegistry().register(ContextSamplingPackage());
 
-    study = Study("1234", "bardram", name: "bardram study")
-      ..dataEndPoint = DataEndPoint(DataEndPointTypes.PRINT)
+    study = Study(id: "1234", userId: "bardram", name: "bardram study")
+      ..dataEndPoint = DataEndPoint(type: DataEndPointTypes.PRINT)
       ..addTriggerTask(
-          ImmediateTrigger(), AutomaticTask(name: 'Task #1')..measures = SamplingSchema.common(namespace: NameSpace.CARP).measures.values.toList());
+          ImmediateTrigger(),
+          AutomaticTask(name: 'Task #1')
+            ..measures = SamplingSchema.common(namespace: NameSpace.CARP)
+                .measures
+                .values
+                .toList());
   });
 
 //  test('Weather', () {
@@ -36,7 +42,8 @@ void main() {
   test('JSON -> Study, assert study id', () async {
     final studyJson = _encode(study);
 
-    Study study_2 = Study.fromJson(json.decode(studyJson) as Map<String, dynamic>);
+    Study study_2 =
+        Study.fromJson(json.decode(studyJson) as Map<String, dynamic>);
     expect(study_2.id, study.id);
 
     print(_encode(study_2));
@@ -45,7 +52,8 @@ void main() {
   test('JSON -> Study, deep assert', () async {
     final studyJson = _encode(study);
 
-    Study study_2 = Study.fromJson(json.decode(studyJson) as Map<String, dynamic>);
+    Study study_2 =
+        Study.fromJson(json.decode(studyJson) as Map<String, dynamic>);
     expect(_encode(study_2), equals(studyJson));
   });
 
@@ -54,12 +62,14 @@ void main() {
     String plainStudyJson = File("test/study_1234.json").readAsStringSync();
     print(plainStudyJson);
 
-    Study plainStudy = Study.fromJson(json.decode(plainStudyJson) as Map<String, dynamic>);
+    Study plainStudy =
+        Study.fromJson(json.decode(plainStudyJson) as Map<String, dynamic>);
     expect(plainStudy.id, study.id);
 
     final studyJson = _encode(study);
 
-    Study study_2 = Study.fromJson(json.decode(plainStudyJson) as Map<String, dynamic>);
+    Study study_2 =
+        Study.fromJson(json.decode(plainStudyJson) as Map<String, dynamic>);
     expect(_encode(study_2), equals(studyJson));
   });
 
@@ -71,7 +81,8 @@ void main() {
     expect(dp_1.header.dataFormat.namespace, NameSpace.CARP);
     print(_encode(dp_1));
 
-    OMHGeopositionDatum geo = TransformerSchemaRegistry().lookup(NameSpace.OMH).transform(loc);
+    OMHGeopositionDatum geo =
+        TransformerSchemaRegistry().lookup(NameSpace.OMH).transform(loc);
     DataPoint dp_2 = DataPoint.fromDatum(study.id, study.userId, geo);
     expect(dp_2.header.dataFormat.namespace, NameSpace.OMH);
     expect(geo.geoposition.latitude.value, loc.latitude);
@@ -84,7 +95,8 @@ void main() {
     expect(dp_1.header.dataFormat.namespace, NameSpace.CARP);
     print(_encode(dp_1));
 
-    OMHPhysicalActivityDatum phy = TransformerSchemaRegistry().lookup(NameSpace.OMH).transform(act);
+    OMHPhysicalActivityDatum phy =
+        TransformerSchemaRegistry().lookup(NameSpace.OMH).transform(act);
     DataPoint dp_2 = DataPoint.fromDatum(study.id, study.userId, phy);
     expect(dp_2.header.dataFormat.namespace, NameSpace.OMH);
     expect(phy.activity.activityName, act.type);
@@ -98,34 +110,46 @@ void main() {
     GeoPosition compute = GeoPosition(55.783499, 12.518914); // DTU Compute
     GeoPosition lyngby = GeoPosition(55.7704, 12.5038); // Kgs. Lyngby
 
-    GeofenceMeasure m = ContextSamplingPackage().common.measures[ContextSamplingPackage.GEOFENCE];
-    Geofence f = Geofence.fromMeasure(m)..dwell = 2 * 1000; // dwell timeout 2 secs.
+    GeofenceMeasure m = ContextSamplingPackage()
+        .common
+        .measures[ContextSamplingPackage.GEOFENCE];
+    Geofence f = Geofence.fromMeasure(m)
+      ..dwell = const Duration(seconds: 2); // dwell timeout 2 secs.
     print(f);
 
     d = f.moved(home);
     print('starting from home - $d');
-    expect(d.type, 'ENTER');
+    expect(d.type, GeofenceType.ENTER);
+
+    d = f.moved(home);
+    print('moved home - $d');
+    expect(d, null);
 
     d = f.moved(dtu);
     print('moved to DTU - $d');
-    expect(d.type, 'ENTER');
-    print(_encode(d));
+    expect(d.type, GeofenceType.EXIT);
+
+    d = f.moved(home);
+    print('moved home - $d');
+    expect(d.type, GeofenceType.ENTER);
 
     d = f.moved(lyngby);
     print('moved to Lyngby - $d');
-    expect(d.type, 'EXIT');
+    expect(d.type, GeofenceType.EXIT);
 
     d = f.moved(compute);
     print('moved to DTU Compute - $d');
-    expect(d.type, 'ENTER');
+    expect(d, null);
 
-    sleep(const Duration(seconds: 3));
-    d = f.moved(dtu);
-    expect(d.type, 'DWELL');
-    print('moved to DTU - $d');
     d = f.moved(home);
-    expect(d.type, 'EXIT');
-    print('went home - $d');
+    print('went home, sleeping - $d');
+    sleep(const Duration(seconds: 3));
+    d = f.moved(home);
+    expect(d.type, GeofenceType.DWELL);
+
+    d = f.moved(compute);
+    print('moved to DTU Compute - $d');
+    expect(d.type, GeofenceType.EXIT);
   });
 
   test('Mobility', () {});